+++ source/Energinet.DataHub.MeteringPoints.Application/Common/ReceiveBusinessRequests/IRequestReceiver.cs
@@ -15,7 +15,7 @@
 using System.Threading.Tasks;
 using Energinet.DataHub.MeteringPoints.Application.MarketDocuments;
 
-namespace Energinet.DataHub.MeteringPoints.Application.CloseDown
+namespace Energinet.DataHub.MeteringPoints.Application.Common.ReceiveBusinessRequests
 {
     /// <summary>
     /// Handles incoming business requests
