+++ source/Energinet.DataHub.MeteringPoints.Application/Common/ReceiveBusinessRequests/RequestReceiver.cs
@@ -18,7 +18,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Energinet.DataHub.MeteringPoints.Application.MarketDocuments;
 
-namespace Energinet.DataHub.MeteringPoints.Application.CloseDown
+namespace Energinet.DataHub.MeteringPoints.Application.Common.ReceiveBusinessRequests
 {
     public class RequestReceiver : IRequestReceiver
     {
